feat(docker): add support for podman and update aliases
This commit adds support for Podman as an alternative to Docker on
Darwin systems. If Podman is installed, it will be used instead of
Docker. The Docker aliases have been updated to use 'docker-compose'
instead of 'docker compose'. Additionally, the Orbstack handling has
been moved to a separate condition to ensure it's only used when
Podman and Colima are not available.